(STRETCH) : refactored code to jsx and folder layout and added styles
